+++ docs/court_wizard.d.ts
@@ -6,15 +6,15 @@ export type InitInput = RequestInfo | URL | Response | BufferSource | WebAssembl
 export interface InitOutput {
   readonly memory: WebAssembly.Memory;
   readonly main: (a: number, b: number) => number;
-  readonly __wasm_bindgen_func_elem_199773: (a: number, b: number, c: number) => void;
-  readonly __wasm_bindgen_func_elem_199560: (a: number, b: number) => void;
-  readonly __wasm_bindgen_func_elem_192937: (a: number, b: number) => void;
-  readonly __wasm_bindgen_func_elem_192948: (a: number, b: number) => void;
-  readonly __wasm_bindgen_func_elem_199759: (a: number, b: number, c: number, d: number) => void;
-  readonly __wasm_bindgen_func_elem_199761: (a: number, b: number, c: number) => void;
-  readonly __wasm_bindgen_func_elem_195062: (a: number, b: number, c: number) => void;
-  readonly __wasm_bindgen_func_elem_195082: (a: number, b: number) => void;
-  readonly __wasm_bindgen_func_elem_199763: (a: number, b: number) => void;
+  readonly __wasm_bindgen_func_elem_199933: (a: number, b: number, c: number) => void;
+  readonly __wasm_bindgen_func_elem_199732: (a: number, b: number) => void;
+  readonly __wasm_bindgen_func_elem_193109: (a: number, b: number) => void;
+  readonly __wasm_bindgen_func_elem_193120: (a: number, b: number) => void;
+  readonly __wasm_bindgen_func_elem_199931: (a: number, b: number, c: number, d: number) => void;
+  readonly __wasm_bindgen_func_elem_199935: (a: number, b: number) => void;
+  readonly __wasm_bindgen_func_elem_199945: (a: number, b: number, c: number) => void;
+  readonly __wasm_bindgen_func_elem_195234: (a: number, b: number, c: number) => void;
+  readonly __wasm_bindgen_func_elem_195254: (a: number, b: number) => void;
   readonly __wbindgen_export: (a: number, b: number) => number;
   readonly __wbindgen_export2: (a: number, b: number, c: number, d: number) => number;
   readonly __wbindgen_export3: (a: number) => void;